requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyJWT==2.10.1
Werkzeug==3.1.3
